==> notebooks/script2_fractionation.ipynb <==
# Python Script 2: Fractionation
# 1) Register 288 new samples
# 2) Register 3 Fraction Plate maps

import pandas as pd
from datetime import date
import numpy as np
import os


# hardcoded variables

CELLULAR_FRACTION = ['Nucleoplasm', 'Chromatin', 'Insoluble'] # iterate through each to get the full 3x fractionation
TX_TIME = 4
TX_TIME_UNIT = "Hours"
VOL = 125
HARVEST_PROTOCOL_ID = "PR003-V1"
FRX_PROTOCOL_ID = "PR004-V1"
PREP_BY = "Bodhi Hueffmeier"

PLATE_BARCODE = "" # see below, this will auto-increment +1 from the input material plate
# TODO do we want the PLATE_BARCODE to be provided by user to ensure no duplicate barcodes?
PLATE_NAME = "Set 1 Rep 1"
WELL_POSITION = "" # see below, inherited from previous plate

# ASSUME TREATMENT PLATE HAS ALREADY BEEN GENERATED PER TOY EXAMPLE BELOW
meta_dir = 'G:/My Drive/Lindsay Pino/proj/2023_scispot_utils/data' # note: link to GDrive location? S3?
treatment_plate_file = os.path.join(meta_dir, "1K THP-1 Screen - Metadata MVP test case - 01-Cell Culture.csv")
treatment_plate = pd.read_csv(treatment_plate_file)
treatment_plate

PLATE_BARCODE = "PL" + str(int(treatment_plate['Plate Barcode'].drop_duplicates()[0][2:]) + 1).rjust(4, '0')

print(PLATE_BARCODE)

# 1) Register 288 new samples
# i.e. 3x 96 samples

def cell_fraction_rows(increment, parent_sample, cell_frx):
    name = ""
    well_position = ""
    
    if parent_sample != "NUC1":
        well_position = str(treatment_plate.loc[treatment_plate['Registry ID'] == parent_sample]['Well Position'].iloc[0])
    
    rows = [
    str("FRA" + str(increment)), # Registry ID - don't do anything with   Registry ID,
    name, # Name (temporary field)                   Name,
    parent_sample, # Parent Sample                   Parent Sample, 
    cell_frx, # Cellular Fraction                    Cellular Fraction,
    TX_TIME, # Post-Treatment Time Point             Post-Treatment Time Point,
    TX_TIME_UNIT, # Time Point Unit                  Time Point Unit
    VOL, # Volume (uL)                               Volume (uL),
    "", # Protein Concentration (ug/uL)              Protein Concentration (ug/uL),
    HARVEST_PROTOCOL_ID, # Harvest Protocol ID       Harvest Protocol ID,
    FRX_PROTOCOL_ID, # Fractionation Protocol ID     Fractionation Protocol ID,
    PREP_BY, # Prepared By                           Prepared By,
    str(date.today()), # Preparation Date            Preparation Date,
    "", # Record Creator                             Record Creator,
    "", # Storage Location                           Storage Location
    PLATE_BARCODE, # Plate Barcode                   Plate Barcode
    str(str(PLATE_NAME) + " " + str(cell_frx)), # Plate Name
    well_position # Well Position
    ]

    return rows

columns = ["Registry ID",
           "Name",
           "Parent Sample",
           "Cellular Fraction",
           "Post-Treatment Time Point",
           "Time Point Unit",
           "Volume (uL)",
           "Protein concentration (ug/ul)",
           "Harvest Protocol ID",
           "Fractionation Protocol ID",
           "Prepared By",
           "Preparation Date",
           "Record Creator",
           "Storage Location",
           "Plate Barcode",
           "Plate Name",
           "Well Position"]

df_out = pd.DataFrame(columns=columns)

i = 1
for cell_frx in CELLULAR_FRACTION:
    new_row = cell_fraction_rows(i, "NUC1", cell_frx) # Generate a row for the 2x cryo nuclei
    new_row[-1] = "A1"
    df_out = df_out.append(pd.DataFrame([new_row], columns=columns)) # add one row for the first cryo nuclei...
    i += 1
    
    new_row[0] =  str("FRA" + str(i))
    new_row[-1] = "A2"
    df_out = df_out.append(pd.DataFrame([new_row], columns=columns)) # ... and a second row for the other cryo nuclei
    i += 1
    
    # now loop through all the "real" samples
    for parent in treatment_plate['Registry ID']: 
        new_row = cell_fraction_rows(i, parent, cell_frx)
        df_out = df_out.append(pd.DataFrame([new_row], columns=columns))
        i += 1
        
    # and finally add the last two wells of NUC1 controls
    new_row = cell_fraction_rows(i, "NUC1", cell_frx) # Generate a row for the 2x cryo nuclei
    new_row[-1] = "H11"
    df_out = df_out.append(pd.DataFrame([new_row], columns=columns)) # add one row for the first cryo nuclei...
    i += 1
    
    new_row[-1] = "H12"
    new_row[0] = str("FRA" + str(i))
    df_out = df_out.append(pd.DataFrame([new_row], columns=columns)) # ... and a second row for the other cryo nuclei
    i += 1
    
df_out

df_out.to_csv(os.path.join(meta_dir, "./script2_fractionation_output.csv"), index=False)
del df_out
