==> notebooks/script3_digestion.ipynb <==
# Python Script 3: Digestion
# 1) Register 96 new samples
# 2) Register 1 Digested Peptides Plate map


import pandas as pd
from datetime import date
import numpy as np
import os


# hardcoded variables

VOL = 170
VOL_POOL = 50
SPIKE_STD = "None"
SPIKE_AMOUNT = ""
SPIKE_LOT = ""
DIGEST_PROTOCOL_ID = "PR005-V1"
PREP_BY = "Julia Robbins"

SET = "1"
REP = "1"

PLATE_BARCODE = "" 
PLATE_NAME = "Set " + SET + " Rep " + REP + " Digested peptides"
WELL_POSITION = ""

# ASSUME FRACTIONATION PLATE HAS ALREADY BEEN GENERATED PER TOY EXAMPLE BELOW
meta_dir = 'G:/My Drive/Lindsay Pino/proj/2023_scispot_utils/data' # note: link to GDrive location? S3?
fractionation_plate_file = os.path.join(meta_dir, "script2_fractionation_output.csv")
fractionation_plate = pd.read_csv(fractionation_plate_file)
fractionation_plate

def peptide_digest_rows(increment, parent_sample, well_position, cell_frx):
    
    name = "lol"
    
    name4plate = PLATE_NAME + " (" + cell_frx + ")"
    
    rows = [
    str("PDG" + str(increment)), # Registry ID - don't do anything with   Registry ID,
    name, # Name (temporary field)                   Name,
    parent_sample, # Parent Sample                   Parent Sample,
    VOL, # Volume (uL)                               Volume (uL),
    "", # Protein Input (ug)                         Protein Input (ug),
    SPIKE_STD, # Spike-In Standard                   Spike-In Standard,
    "", # Spike-In Amount (ng)                       Spike-In Amount (ng)
    "", # Spike-In Lot ID                            Spike-In Lot ID 
    DIGEST_PROTOCOL_ID, # Digest Protocol ID         Digest Protocol ID,
    PREP_BY, # Prepared By                           Prepared By,
    str(date.today()), # Preparation Date            Preparation Date,
    "", # Record Creator                             Record Creator,
    "", # Storage Location                           Storage Location
    "", # Plate Barcode
    name4plate, # Plate Name
    well_position # Well position
    ]

    return rows

def peptide_pool_rows(increment, parent_sample, well_position):
    
    name = "lol"
    
    name4plate = str(PLATE_NAME + " (nuclear, dilute)")
    
    rows = [
    str("PDG" + str(increment)), # Registry ID - don't do anything with   Registry ID,
    name, # Name (temporary field)                   Name,
    parent_sample, # Parent Sample                   Parent Sample,
    VOL_POOL, # Volume (uL)                               Volume (uL),
    "", # Protein Input (ug)                         Protein Input (ug),
    "", # Spike-In Standard                   Spike-In Standard,
    "", # Spike-In Amount (ng)                       Spike-In Amount (ng)
    "", # Spike-In Lot ID                            Spike-In Lot ID 
    "", # Digest Protocol ID         Digest Protocol ID,
    PREP_BY, # Prepared By                           Prepared By,
    str(date.today()), # Preparation Date            Preparation Date,
    "", # Record Creator                             Record Creator,
    "", # Storage Location                            Storage Location
    "", # Plate Barcode
    name4plate, # Plate Name
    well_position # Well position
    ]

    return rows

columns = [
    "Registry ID", 
    "Name", 
    "Parent Sample",
    "Volume (uL)", 
    "Protein Input (ug)", 
    "Spike-In Standard", 
    "Spike-In Amount", 
    "Spike-In Lot ID", 
    "Digest Protocol ID", 
    "Prepared By", 
    "Preparation Date", 
    "Record Creator", 
    "Storage Location", 
    "Plate Barcode", 
    "Plate Name",
    "Well Position"
    ]

df_out = pd.DataFrame(columns=columns)
i = 1
if REP == "1":
    
    fractionation_plate = fractionation_plate[(fractionation_plate['Cellular Fraction'] == 'Nucleoplasm') |
                                              (fractionation_plate['Cellular Fraction'] == 'Chromatin')]
    
    # now loop through all the "real" samples
    for parent in fractionation_plate['Registry ID']: 
        well_pos = str(fractionation_plate.loc[fractionation_plate['Registry ID'] == parent]['Well Position'].iloc[0])
        cell_fraction = str(fractionation_plate.loc[fractionation_plate['Registry ID'] == parent]['Cellular Fraction'].iloc[0])
        new_row = peptide_digest_rows(i, parent, well_pos, cell_fraction)
        df_out = df_out.append(pd.DataFrame([new_row], columns=columns))
        i += 1
    
    # for Rep 1, we'll also be pooling nuc+chrom for DDA
    # so while we're making these digests, also make the bonus pooled plate
    df_out_pool = pd.DataFrame(columns = columns) 
    
    # hard code NUC1 by plate positions for A1 and A2 plate positions
    temp = fractionation_plate[fractionation_plate['Well Position'] == "A1"]
    well_pos = str(temp['Well Position'].drop_duplicates().iloc[0])
    related_fra = str(', '.join(list(temp['Registry ID'])))
    new_row = peptide_pool_rows(i, related_fra, well_pos)
    df_out_pool = df_out_pool.append(pd.DataFrame([new_row], columns=columns))
    i += 1
    
    temp = fractionation_plate[fractionation_plate['Well Position'] == "A2"]
    well_pos = str(temp['Well Position'].drop_duplicates().iloc[0])
    related_fra = str(', '.join(list(temp['Registry ID'])))
    new_row = peptide_pool_rows(i, related_fra, well_pos)
    df_out_pool = df_out_pool.append(pd.DataFrame([new_row], columns=columns))
    i += 1
    
    for grandparent in fractionation_plate['Parent Sample'].drop_duplicates():
        if grandparent != "NUC1":
            temp = fractionation_plate[fractionation_plate['Parent Sample'] == grandparent]
            well_pos = str(temp['Well Position'].drop_duplicates().iloc[0])
            related_fra = str(', '.join(list(temp['Registry ID'])))
            new_row = peptide_pool_rows(i, related_fra, well_pos)
            df_out_pool = df_out_pool.append(pd.DataFrame([new_row], columns=columns))
            i += 1
        # hard code NUC1 here?
    
    # hard code NUC1 by plate positions for H11 and H12 plate positions
    temp = fractionation_plate[fractionation_plate['Well Position'] == "H11"]
    well_pos = str(temp['Well Position'].drop_duplicates().iloc[0])
    related_fra = str(', '.join(list(temp['Registry ID'])))
    new_row = peptide_pool_rows(i, related_fra, well_pos)
    df_out_pool = df_out_pool.append(pd.DataFrame([new_row], columns=columns))
    i += 1
    
    temp = fractionation_plate[fractionation_plate['Well Position'] == "H12"]
    well_pos = str(temp['Well Position'].drop_duplicates().iloc[0])
    related_fra = str(', '.join(list(temp['Registry ID'])))
    new_row = peptide_pool_rows(i, related_fra, well_pos)
    df_out_pool = df_out_pool.append(pd.DataFrame([new_row], columns=columns))
    i += 1

else:
    
    fractionation_plate = fractionation_plate[(fractionation_plate['Cellular Fraction'] == 'Chromatin')]

    # loop through all the chromatin samples from the fractionation plate to create digests
    for parent in fractionation_plate['Registry ID']: 
        well_pos = str(fractionation_plate.loc[fractionation_plate['Registry ID'] == parent]['Well Position'].iloc[0])
        cell_fraction = "Chromatin"
        new_row = peptide_digest_rows(i, parent, well_pos, cell_fraction)
        df_out = df_out.append(pd.DataFrame([new_row], columns=columns))
        i += 1

#print(df_out.head())
#print(df_out_pool.head())

df_out.to_csv(os.path.join(meta_dir, "./script3_digestion_output_rep1.csv"), index=False)
df_out_pool.to_csv(os.path.join(meta_dir, "./script4b_digestion_pools_output_rep1.csv"), index=False)
del df_out
del df_out_pool
