==> notebooks/script5_ms.ipynb <==
# Python Script 5: MS runs

import pandas as pd
from datetime import date
import numpy as np
import os


# hardcoded variables

PREP_BY = "Andrea Gutierrez"
ACQUISITION_TYPE = "DDA" # DDA OR DIA
GRAD_LENGTH = "44"
INSTRUMENT = "timsTOF SCP"
INSTRUMENT_LC = "Evosep"
LC_DECK_POSITION = ""
COLUMN_ID = "MSC1"
INJECT_VOL = "26"
EVOTIP_LOT_NUM = "AB1134"
SAMPLE_LOAD_PROTOCOL = "PR008-V1"

SET = "1"
REP = "1"

PLATE_BARCODE = "" 
PLATE_NAME = "Set " + SET + " Rep " + REP + ACQUISITION_TYPE + " Evotips"
WELL_POSITION = ""

# ASSUME DIGESTION PLATE HAS ALREADY BEEN GENERATED PER TOY EXAMPLE BELOW
meta_dir = 'G:/My Drive/Lindsay Pino/proj/2023_scispot_utils/data' # note: link to GDrive location? S3?
digestion_plate_file = os.path.join(meta_dir, "script4b_digestion_pools_output_rep1.csv")
digestion_plate = pd.read_csv(digestion_plate_file)
digestion_plate

def ms_run_rows(increment, parent_sample, well_position):
    
    name = "lol"
    
    rows = [
    str("MSR" + str(increment)), # "Registry ID"
    name, # "Run Name",
    str(date.today()), # "Run Date",
    PREP_BY, # "Prepared By",
    ACQUISITION_TYPE, # "Acquisition Type",
    GRAD_LENGTH, # "Gradient Length (Minutes)",
    INSTRUMENT, #"Instrument",
    INSTRUMENT_LC, #"Instrument LC Type",
    LC_DECK_POSITION, # "LC Deck Position",
    COLUMN_ID, # "Column ID"
    parent_sample, # "Peptide Digest IDs",
    INJECT_VOL, # "Injection Volume (uL)",
    "", # "Injection Mass",
    "", # "Injection Mass Unit",
    EVOTIP_LOT_NUM, # "Evotip Lot Number",
    SAMPLE_LOAD_PROTOCOL, # "Sample Loading Protocol ID",
    "", # "MS Acquisition Protocol ID",
    "", # "Record Creator",
    "", # "Evotip Rack Barcode",
    "", # "Evotip Rack Name"
    well_position #"Tip Position"
    ]

    return rows

columns = [
    "Registry ID",
    "Run Name",
    "Run Date",
    "Prepared By",
    "Acquisition Type",
    "Gradient Length (Minutes)",
    "Instrument",
    "Instrument LC Type",
    "LC Deck Position",
    "Column ID",
    "Peptide Digest IDs",
    "Injection Volume (uL)",
    "Injection Mass",
    "Injection Mass Unit",
    "Evotip Lot Number",
    "Sample Loading Protocol ID",
    "MS Acquisition Protocol ID",
    "Record Creator",
    "Evotip Rack Barcode",
    "Evotip Rack Name",
    "Tip Position"
]

df_out = pd.DataFrame(columns=columns)

i = 1
for parent in digestion_plate['Registry ID']: 
    well_pos = str(digestion_plate.loc[digestion_plate['Registry ID'] == parent]['Well Position'].iloc[0])
    new_row = ms_run_rows(i, parent, well_pos)
    df_out = df_out.append(pd.DataFrame([new_row], columns=columns))
    i += 1

#df_out

df_out.to_csv(os.path.join(meta_dir, "./script5_msrun_output.csv"), index=False)
#del df_out
